requirements.txt
cachetools==4.2.4
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.69.1
grpcio==1.71.0rc2
grpcio-status==1.49.0rc1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
requests==2.32.3
rich==13.3.5
rsa==4.9
setuptools==75.8.2
six==1.17.0
urllib3==2.3.0
python-dotenv==1.0.0
google-api-python-client==2.97.0
